Move Patir::CommandSequence impl and test in file
This shall increase the code clarity.
